Fix: normalize Steam ID to 64-bit and use BigIntegerField
- Accept both 32-bit (account_id) and 64-bit (friend_id) formats in profile update
- Auto-convert 32-bit to 64-bit for consistent storage
- Change steamid field from IntegerField to BigIntegerField

+++ backend/app/models.py
@@ -70,7 +70,7 @@ class DraftStyles(StrEnum):
 
 
 class CustomUser(AbstractUser):
-    steamid = models.IntegerField(null=True, unique=True, blank=True)
+    steamid = models.BigIntegerField(null=True, unique=True, blank=True)
     nickname = models.TextField(null=True, blank=True)
     mmr = models.IntegerField(null=True, blank=True)
     league_mmr = models.IntegerField(null=True, blank=True)
